Remove unnecessary construction
