+++ src/go/svcfs/json_meta_data_repo.go
@@ -1,56 +1,23 @@
 package svcfs
 
 import (
-	"errors"
 	"fmt"
-	"io/fs"
 	"net/url"
-	"os"
 	"path/filepath"
 
 	core "github.com/kkrull/marmot/corerepository"
 )
 
-// A meta repo that stores meta data in JSON files in the specified directory.
 func NewJsonMetaDataRepo(repositoryPath string) *JsonMetaDataRepo {
 	metaDataDir := filepath.Join(repositoryPath, ".marmot")
 	return &JsonMetaDataRepo{
-		repositoryDir: repositoryPath,
-		metaDataFile:  filepath.Join(metaDataDir, "meta-repo.json"),
+		metaDataFile: filepath.Join(metaDataDir, "meta-repo.json"),
 	}
 }
 
+// A meta repo that stores meta data in JSON files in the specified directory.
 type JsonMetaDataRepo struct {
-	// metaDataDir   string
-	metaDataFile  string
-	repositoryDir string
-}
-
-/* MetaDataAdmin */
-
-func (*JsonMetaDataRepo) Create(repositoryDir string) error {
-	_, statErr := os.Stat(repositoryDir)
-	if errors.Is(statErr, fs.ErrNotExist) {
-		return initDirectory(repositoryDir)
-	} else if statErr != nil {
-		return fmt.Errorf("failed to check for existing meta repo %s; %w", repositoryDir, statErr)
-	} else {
-		return fmt.Errorf("path already exists: %s", repositoryDir)
-	}
-}
-
-func initDirectory(repositoryDir string) error {
-	metaDataDir := filepath.Join(repositoryDir, ".marmot")
-	metaDataFile := filepath.Join(metaDataDir, "meta-repo.json")
-
-	emptyFile := EmptyMetaRepoFile("0.0.1")
-	if dirErr := os.MkdirAll(metaDataDir, fs.ModePerm); dirErr != nil {
-		return fmt.Errorf("failed to make directory %s; %w", metaDataDir, dirErr)
-	} else if writeErr := emptyFile.WriteTo(metaDataFile); writeErr != nil {
-		return fmt.Errorf("failed to write file %s; %w", metaDataFile, writeErr)
-	} else {
-		return nil
-	}
+	metaDataFile string
 }
 
 /* RepositorySource */
